refactor(pipeline): environment loader module
- [x] a separate module for the data pipeline environment loader;
