+++ src/data_pipeline/mod.rs
@@ -1,23 +1,28 @@
+/// Data pipeline module.
+///
+/// Environment variables should be placed in the `.env` file:
+///
+/// GITHUB_TOKEN=...
+/// GPG_PASSPHRASE=...
+/// MONGODB_CONNECTION_STRING=...
+/// MONGODB_DATABASE=...
+///
 use colored::Colorize;
 use octorust::{
-    auth::Credentials,
     types::SearchReposSort,
     types::{Order, RepoSearchResultItem},
-    Client,
 };
 use std::{
     cmp::Ordering,
-    collections::HashMap,
     env::{self, args, Args},
     fs::{self, File},
     io,
-    process::Command,
 };
 
 mod artifact;
+mod environment;
+mod github;
 mod mongo;
-
-type Result<T> = std::result::Result<T, Box<dyn std::error::Error + Send + Sync>>;
 
 /// The entry point of the program.
 pub fn main() {
@@ -29,13 +34,6 @@ struct InuputArguments {
     context: Option<String>,
     collection: Option<String>,
     search_term: Option<String>,
-}
-
-/// GitHub repos fetch result.
-struct FetchResult {
-    items: Vec<RepoSearchResultItem>,
-    total: i64,
-    retry: bool,
 }
 
 /// Supported contexts.
@@ -50,7 +48,7 @@ struct DataPipeline<'a> {
 }
 
 impl<'a> DataPipeline<'a> {
-    /// Creates a new data pipeline.
+    /// Program constructor.
     fn new() -> DataPipeline<'a> {
         let contexts: Contexts = ["Create artifact", "Restore artifact"];
         let collections: Collections = ["repos"];
@@ -66,7 +64,7 @@ impl<'a> DataPipeline<'a> {
     fn init(&mut self) {
         println!("\n{}", "DataPipeline initialized.".blue().bold());
 
-        let _result = self.load_env_vars();
+        environment::main(None);
 
         let args = self.args();
 
@@ -99,7 +97,7 @@ impl<'a> DataPipeline<'a> {
         }
     }
 
-    /// Parses the data pipeline arguments.
+    /// Parses arguments passed to the program.
     fn args(&mut self) -> InuputArguments {
         let mut args: Args = args();
 
@@ -152,7 +150,7 @@ impl<'a> DataPipeline<'a> {
         }
     }
 
-    /// Prints the context selection instructions.
+    /// Prints instructions for selecting a context.
     fn print_context_instructions(&self) {
         println!("\n{}", "Available contexts:".yellow().bold());
 
@@ -217,7 +215,7 @@ impl<'a> DataPipeline<'a> {
         }
     }
 
-    /// Prints the collection selection instructions.
+    /// Prints instructions for selecting a collection.
     fn print_collection_instructions(&self) {
         println!("\n{}", "Available collections:".yellow().bold());
 
@@ -242,9 +240,9 @@ impl<'a> DataPipeline<'a> {
         collection_index
     }
 
-    /// Resets the input argument to start over if the program does not exist.
+    /// Resets the input argument to start over if the option does not exist.
     fn reset_input_arg(&self) -> String {
-        println!("\n{}", "The subprogram does not exist.".red());
+        println!("\n{}", "Invalid option.".red());
         String::new()
     }
 
@@ -286,7 +284,7 @@ impl<'a> DataPipeline<'a> {
             let per_page = 5;
             page += 1;
 
-            let mut fetch_result = FetchResult {
+            let mut fetch_result = github::FetchResult {
                 items: Vec::<RepoSearchResultItem>::new(),
                 total: 0,
                 retry: false,
@@ -298,7 +296,7 @@ impl<'a> DataPipeline<'a> {
                 .unwrap();
 
             runtime.block_on(async {
-                let result = self.repos_request(q, sort, order, per_page, page).await;
+                let result = github::repos(q, sort, order, per_page, page).await;
                 fetch_result = match result {
                     Ok(data) => {
                         if data.retry {
@@ -321,7 +319,7 @@ impl<'a> DataPipeline<'a> {
                         println!("\n{}: {:?}", "There was an error".red(), error);
                         let items = Vec::<RepoSearchResultItem>::new();
                         let total: i64 = 0;
-                        FetchResult {
+                        github::FetchResult {
                             items,
                             total,
                             retry: false,
@@ -347,188 +345,5 @@ impl<'a> DataPipeline<'a> {
                 break;
             }
         }
-    }
-
-    /// GitHub repositories request.
-    async fn repos_request(
-        &self,
-        q: &str,
-        sort: SearchReposSort,
-        order: Order,
-        per_page: i64,
-        page: i64,
-    ) -> Result<FetchResult> {
-        let token_env = env::var("GITHUB_TOKEN");
-        let token = match token_env.unwrap().trim().parse::<String>() {
-            Ok(value) => value,
-            Err(_) => String::new(),
-        };
-
-        let github = Client::new(String::from("user-agent-name"), Credentials::Token(token));
-
-        let mut retry: bool = false;
-
-        let client = github.unwrap();
-        let search = client.search();
-        let result = search.repos(q, sort, order, per_page, page).await;
-        let raw_res = match result {
-            Ok(res) => Ok(res),
-            Err(error) => {
-                println!(
-                    "\n{}: {:?}",
-                    "There was an error getting data from GitHub".red(),
-                    error
-                );
-
-                let err = error.to_string();
-
-                let rate_limit_regx =
-                    regex::Regex::new(r"(Rate limited for the next)\s+(\d+)\s+(seconds)").unwrap();
-                let captures = rate_limit_regx.captures(&err).map(|captures| {
-                    captures
-                        .iter() // All the captured groups
-                        .skip(1) // Skipping the complete match
-                        // .flat_map(|c| c) // Ignoring all empty optional matches
-                        .flatten()
-                        .map(|c| c.as_str()) // Grab the original strings
-                        .collect::<Vec<_>>() // Create a vector
-                });
-                let wait_timeout = match captures.as_deref() {
-                    Some(["Rate limited for the next", x, "seconds"]) => {
-                        let x: i64 = x.parse().expect("Can't parse number");
-                        x
-                    }
-                    _ => panic!("Unknown Command: {}", &err),
-                };
-                println!(
-                    "\n{}: {:?}",
-                    "GitHub API rate limit hit. Will wait for".red(),
-                    wait_timeout
-                );
-
-                retry = true;
-
-                let mut child = Command::new("sleep")
-                    .arg(wait_timeout.to_string())
-                    .spawn()
-                    .unwrap();
-                let _result = child.wait().unwrap();
-
-                Err(())
-            }
-        };
-
-        if retry {
-            let result = FetchResult {
-                items: Vec::<RepoSearchResultItem>::new(),
-                total: 0,
-                retry: true,
-            };
-            return Ok(result);
-        }
-
-        let res = raw_res.unwrap();
-        let body = res.body.to_owned();
-        let items = body.items;
-        for item in items {
-            let name = item.full_name;
-            println!("\n{}: {}", "Data".cyan(), name);
-        }
-
-        println!("{}: {}", "Response".green(), res.status);
-        println!("{}: {:#?}\n", "Headers".green(), res.headers);
-        println!("{}: {:#?}\n", "Body".green(), res.body);
-
-        println!("\n\n{}", "Done!".green().bold());
-
-        let items = res.body.items.to_owned();
-        let total = res.body.total_count.to_owned();
-
-        let result = FetchResult {
-            items,
-            total,
-            retry: false,
-        };
-        Ok(result)
-    }
-
-    /// Load environment variables read from the .env file.
-    fn load_env_vars(&self) -> HashMap<String, String> {
-        let result = self.read_env();
-        let config = match result {
-            Ok(env) => {
-                let iter = env.iter();
-                for record in iter {
-                    let key = record.0;
-                    let value = record.1;
-                    env::set_var(key, value);
-                    println!("key: {key}\nvalue: {value}");
-                }
-                env
-            }
-            Err(error) => {
-                println!(
-                    "\n{}: {:?}",
-                    "There was an error reading the .env file".red(),
-                    error
-                );
-                HashMap::from([])
-            }
-        };
-        config
-    }
-
-    /// Read variables from the .env file.
-    fn read_env(&self) -> std::io::Result<HashMap<String, String>> {
-        let cwd = env::current_dir()?;
-
-        println!("The current directory is {}", cwd.display());
-
-        let env_path = cwd.display().to_string() + "/.env";
-        let env_path_str = env_path.as_str();
-        let contents = fs::read_to_string(env_path_str).expect("Can't read file");
-
-        println!("Text content:\n{contents}");
-
-        let lines = self.read_lines(env_path_str);
-        let config: Vec<_> = lines
-            .iter()
-            .flat_map(|x| {
-                let split_pair = x.split('=').collect::<Vec<&str>>();
-                let split_key = split_pair.first();
-                let some_key = split_key.is_some();
-                let key = if some_key {
-                    match split_key.unwrap().trim().parse::<String>() {
-                        Ok(value) => value,
-                        Err(_) => String::new(),
-                    }
-                } else {
-                    String::new()
-                };
-                let split_value = split_pair.get(1);
-                let some_value = split_value.is_some();
-                let value = if some_value {
-                    match split_value.unwrap().trim().parse::<String>() {
-                        Ok(value) => value,
-                        Err(_) => String::new(),
-                    }
-                } else {
-                    String::new()
-                };
-                let map = vec![(key, value)];
-                map
-            })
-            .collect();
-        let map: HashMap<String, String> = HashMap::from_iter(config);
-        Ok(map)
-    }
-
-    /// Read a file line by line.
-    fn read_lines(&self, filename: &str) -> Vec<String> {
-        let mut result = Vec::new();
-        for line in fs::read_to_string(filename).unwrap().lines() {
-            result.push(line.to_string())
-        }
-        result
     }
 }
